be/requirements.txt
uvicorn==0.24.0.post1
fastapi==0.104.1
pydantic==2.5.2
cassandra-driver==3.28.0
pydantic-settings==2.1.0
python-dotenv~=1.0.0
contextvars==2.4
pyjwt==2.8.0
